Rename ifsbench.cli to ifsbench.command_line to avoid name clashes
